refactor: extract system health and requeue helpers to system_health.py
Move SYSTEMIC_FAILURE_THRESHOLD, _get_system_health_path, _load_system_health,
_save_system_health, _record_systemic_failure, reset_systemic_failures,
_spawn_diagnostic_agent, _auto_pause_and_diagnose, _handle_systemic_failure,
_requeue_task, _requeue_task_blameless, and _reset_systemic_failure_counter
(~300 lines) out of scheduler.py into system_health.py.

Also deduplicates the three system health path/load/save functions that were
defined twice in scheduler.py. The canonical (second) definition is kept,
which includes auto_paused/auto_pause_reason fields in defaults.

scheduler.py re-imports and re-exports all moved names to preserve backwards
compatibility for tests and jobs.py.
