+++ octopoid/scheduler.py
@@ -67,6 +67,20 @@ from .prompt_renderer import (
     _load_intervention_context_for_prompt,
     _render_prompt,
 )
+from .system_health import (
+    SYSTEMIC_FAILURE_THRESHOLD,
+    _get_system_health_path,
+    _load_system_health,
+    _save_system_health,
+    _record_systemic_failure,
+    reset_systemic_failures,
+    _spawn_diagnostic_agent,
+    _auto_pause_and_diagnose,
+    _handle_systemic_failure,
+    _requeue_task,
+    _requeue_task_blameless,
+    _reset_systemic_failure_counter,
+)
 
 logger = logging.getLogger("octopoid.scheduler")
 
@@ -2122,314 +2136,8 @@ def run_housekeeping() -> None:
             logger.debug(f"Housekeeping job {job.__name__} failed: {e}")
 
 
-# =============================================================================
-# System Health and Auto-Pause
-# =============================================================================
-
-SYSTEMIC_FAILURE_THRESHOLD = 2
-
-
-def _get_system_health_path() -> Path:
-    """Get path to the system health tracking file."""
-    from .config import get_runtime_dir
-    return get_runtime_dir() / "system_health.json"
-
-
-def _load_system_health() -> dict:
-    """Load system health state from disk.
-
-    Returns:
-        Dict with consecutive_systemic_failures, last_failure_time,
-        last_failure_reason, last_diagnostic_spawned.
-    """
-    path = _get_system_health_path()
-    if path.exists():
-        try:
-            return json.loads(path.read_text())
-        except (json.JSONDecodeError, OSError):
-            pass
-    return {
-        "consecutive_systemic_failures": 0,
-        "last_failure_time": None,
-        "last_failure_reason": None,
-        "last_diagnostic_spawned": None,
-    }
-
-
-def _save_system_health(health: dict) -> None:
-    """Persist system health state to disk."""
-    path = _get_system_health_path()
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(health, indent=2))
-
-
-def _record_systemic_failure(reason: str) -> int:
-    """Increment the consecutive systemic failure counter.
-
-    Args:
-        reason: Human-readable description of the failure.
-
-    Returns:
-        The new consecutive failure count.
-    """
-    health = _load_system_health()
-    health["consecutive_systemic_failures"] = health.get("consecutive_systemic_failures", 0) + 1
-    health["last_failure_time"] = datetime.now(tz=timezone.utc).isoformat()
-    health["last_failure_reason"] = reason
-    _save_system_health(health)
-    count = health["consecutive_systemic_failures"]
-    logger.warning(f"Systemic failure #{count}: {reason}")
-    return count
-
-
-def reset_systemic_failures() -> None:
-    """Reset the systemic failure counter (called by the diagnostic agent after fixing)."""
-    health = _load_system_health()
-    health["consecutive_systemic_failures"] = 0
-    health["last_failure_time"] = None
-    health["last_failure_reason"] = None
-    _save_system_health(health)
-    logger.info("Systemic failure counter reset")
-
-
-def _spawn_diagnostic_agent(reason: str) -> None:
-    """Spawn the diagnostic agent to investigate a systemic pause.
-
-    Prepares a job directory for the diagnostic agent with the failure
-    context written to context.json, then invokes claude directly.
-
-    Args:
-        reason: The failure reason that triggered the auto-pause.
-    """
-    import yaml as _yaml
-
-    agents_base = get_agents_base_dir()
-    diag_agent_dir = agents_base / "diagnostic"
-    agent_yaml = diag_agent_dir / "agent.yaml"
-
-    if not agent_yaml.exists():
-        logger.error("Diagnostic agent config not found at %s, cannot spawn", agent_yaml)
-        return
-
-    with open(agent_yaml) as f:
-        agent_config = _yaml.safe_load(f) or {}
-
-    agent_config["agent_dir"] = str(diag_agent_dir)
-    agent_config.setdefault("name", "diagnostic")
-    agent_config.setdefault("blueprint_name", "diagnostic")
-
-    # Don't spawn if already running
-    if count_running_instances("diagnostic") > 0:
-        logger.info("Diagnostic agent already running, skipping spawn")
-        return
-
-    try:
-        job_dir = prepare_job_directory("diagnostic", agent_config)
-    except Exception as e:
-        logger.error(f"Failed to prepare diagnostic job directory: {e}")
-        return
-
-    # Read last N lines of scheduler log for context
-    log_tail_lines: list[str] = []
-    log_file = get_logs_dir() / "octopoid.log"
-    if log_file.exists():
-        try:
-            lines = log_file.read_text().splitlines()
-            log_tail_lines = lines[-100:]
-        except OSError:
-            pass
-
-    # Write diagnostic context to job_dir/context.json
-    health = _load_system_health()
-    context = {
-        "trigger_reason": reason,
-        "consecutive_failures": health.get("consecutive_systemic_failures", 0),
-        "last_failure_time": health.get("last_failure_time"),
-        "orchestrator_dir": str(get_orchestrator_dir()),
-        "pause_file": str(get_orchestrator_dir() / "PAUSE"),
-        "health_file": str(_get_system_health_path()),
-        "log_file": str(log_file),
-        "log_tail": log_tail_lines,
-        "queue_counts": None,
-    }
-
-    # Try to get queue counts from the server
-    try:
-        orch_id = queue_utils.get_orchestrator_id()
-        sdk = queue_utils.get_sdk()
-        poll_data = sdk.poll(orch_id)
-        context["queue_counts"] = poll_data.get("queue_counts")
-    except Exception:
-        pass  # Server may be unreachable — that could be the problem
-
-    (job_dir / "context.json").write_text(json.dumps(context, indent=2))
-
-    try:
-        pid = invoke_claude(job_dir, agent_config)
-        register_instance_pid("diagnostic", pid, "", "diagnostic-1")
-        health["last_diagnostic_spawned"] = datetime.now(tz=timezone.utc).isoformat()
-        _save_system_health(health)
-        logger.info(f"Diagnostic agent spawned with PID {pid}")
-    except Exception as e:
-        logger.error(f"Failed to spawn diagnostic agent: {e}")
-
-
-def _auto_pause_and_diagnose(reason: str) -> None:
-    """Write the PAUSE file and spawn the diagnostic agent.
-
-    Called when the systemic failure counter reaches the threshold.
-
-    Args:
-        reason: The failure reason that triggered the auto-pause.
-    """
-    pause_file = get_orchestrator_dir() / "PAUSE"
-    pause_file.write_text(f"Auto-paused: {reason}\n")
-    logger.warning(
-        "System auto-paused due to %d consecutive systemic failures. "
-        "Remove .octopoid/PAUSE when resolved.",
-        SYSTEMIC_FAILURE_THRESHOLD,
-    )
-    _spawn_diagnostic_agent(reason)
-
-
-def _handle_systemic_failure(reason: str) -> None:
-    """Record a systemic failure and auto-pause if the threshold is reached.
-
-    Called from spawn failure handlers and other systemic error paths.
-
-    Args:
-        reason: Human-readable description of what failed.
-    """
-    count = _record_systemic_failure(reason)
-    if count >= SYSTEMIC_FAILURE_THRESHOLD:
-        _auto_pause_and_diagnose(reason)
-
-
-# =============================================================================
-# Spawn Strategies
-# =============================================================================
-
-def _requeue_task(task_id: str, source_queue: str = "incoming", task: dict | None = None) -> None:
-    """Requeue a task back to its source queue after spawn failure.
-
-    Increments attempt_count on the server. If attempt_count reaches the
-    circuit breaker threshold, moves the task to failed instead of requeuing.
-
-    Args:
-        task_id: Task to requeue.
-        source_queue: Queue the task should return to. For tasks claimed from
-            'incoming' (implementer), this is 'incoming'. For tasks claimed
-            in-place from 'provisional' (gatekeeper), this is 'provisional'.
-        task: Optional task dict with current state (used to read attempt_count
-            without an extra API call).
-    """
-    try:
-        from .queue_utils import get_sdk
-        sdk = get_sdk()
-
-        # Circuit breaker only applies to claimed→incoming transitions
-        if source_queue == "incoming":
-            if task is None:
-                task = sdk.tasks.get(task_id) or {}
-            current_attempt_count = task.get("attempt_count", 0)
-            new_attempt_count = current_attempt_count + 1
-            threshold = _get_circuit_breaker_threshold()
-
-            if new_attempt_count >= threshold:
-                reason = (
-                    f"Circuit breaker: spawn failed {new_attempt_count} time(s) "
-                    f"(threshold={threshold}). Task could not be started."
-                )
-                queue_utils.fail_task(
-                    task_id,
-                    reason=reason,
-                    source='spawn-failure-circuit-breaker',
-                    claimed_by=None,
-                    lease_expires_at=None,
-                    attempt_count=new_attempt_count,
-                )
-                logger.debug(f"Circuit breaker tripped for {task_id}: {reason}")
-                return
-
-            sdk.tasks.update(
-                task_id,
-                queue=source_queue,
-                claimed_by=None,
-                lease_expires_at=None,
-                attempt_count=new_attempt_count,
-            )
-        else:
-            # Provisional: just clear the claim, no attempt_count increment
-            sdk.tasks.update(task_id, queue=source_queue, claimed_by=None, lease_expires_at=None)
-
-        logger.debug(f"Requeued task {task_id} back to {source_queue}")
-    except Exception as e:
-        logger.error(f"Requeue failed for {task_id}: {e}")
-
-
-def _get_system_health_path() -> Path:
-    """Get path to system_health.json in the runtime directory."""
-    from .config import get_runtime_dir
-    return get_runtime_dir() / "system_health.json"
-
-
-def _load_system_health() -> dict:
-    """Load system health state, returning defaults if file doesn't exist."""
-    path = _get_system_health_path()
-    if not path.exists():
-        return {
-            "consecutive_systemic_failures": 0,
-            "last_systemic_failure": None,
-            "auto_paused": False,
-            "auto_paused_at": None,
-            "auto_pause_reason": None,
-        }
-    try:
-        return json.loads(path.read_text())
-    except Exception:
-        return {
-            "consecutive_systemic_failures": 0,
-            "last_systemic_failure": None,
-            "auto_paused": False,
-            "auto_paused_at": None,
-            "auto_pause_reason": None,
-        }
-
-
-def _save_system_health(data: dict) -> None:
-    """Persist system health state to disk."""
-    path = _get_system_health_path()
-    path.parent.mkdir(parents=True, exist_ok=True)
-    path.write_text(json.dumps(data, indent=2))
-
-
-def _requeue_task_blameless(task_id: str, source_queue: str = "incoming") -> None:
-    """Requeue a task without incrementing attempt_count.
-
-    Used for systemic failures (spawn failure, infrastructure error) where the
-    task itself is not at fault. The task is returned to its source queue with
-    no penalty.
-
-    Args:
-        task_id: Task to requeue.
-        source_queue: Queue the task should return to.
-    """
-    try:
-        from .queue_utils import get_sdk
-        sdk = get_sdk()
-        sdk.tasks.update(task_id, queue=source_queue, claimed_by=None, lease_expires_at=None)
-        logger.debug(f"Blameless requeue of task {task_id} back to {source_queue} (no attempt_count increment)")
-    except Exception as e:
-        logger.error(f"Blameless requeue failed for {task_id}: {e}")
-
-
-def _reset_systemic_failure_counter() -> None:
-    """Reset the systemic failure counter after a successful spawn."""
-    health = _load_system_health()
-    if health.get("consecutive_systemic_failures", 0) > 0:
-        health["consecutive_systemic_failures"] = 0
-        _save_system_health(health)
-        logger.debug("Systemic failure counter reset after successful spawn")
+# System health, auto-pause, and requeue functions live in system_health.py
+# (imported at the top of this module)
 
 
 def _init_submodule(agent_name: str) -> None:
